+++ priv/javascript/elixir.js
@@ -16,3 +16,4 @@ export { default as Agent } from './lib/agent';
 export { default as Keyword } from './lib/keyword';
 export { default as Base } from './lib/base';
 export { default as String } from './lib/string';
+export { default as Bitwise } from './lib/bitwise';