+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Enhancements
   * Added Base module with function: encode64, decode64, and decode64!
   * Added String module
+  * Added Bitwise module
 
 # v0.12.0
 * Enhancements